refactor code out of main plugin file
